# api/models/event_object.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
)

// EventObject event object
//
// swagger:model EventObject
type EventObject map[string]map[string]string

// Validate validates this event object
func (m EventObject) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this event object based on context it is used
func (m EventObject) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}