# api/restapi/api.go
// Code generated by gin-swagger; DO NOT EDIT.

package restapi

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"
	"net/http"
	"os"
	"os/signal"
	"strings"
	"syscall"
	"time"

	"github.com/gin-contrib/pprof"
	"github.com/gin-gonic/gin"
	"github.com/mikkeloscar/gin-swagger/api"
	"github.com/mikkeloscar/gin-swagger/middleware"
	"github.com/mikkeloscar/gin-swagger/tracing"
	opentracing "github.com/opentracing/opentracing-go"
	log "github.com/sirupsen/logrus"

	operations "github.com/supby/zigbee2webthings/api/restapi/operations"
)

// Routes defines all the routes of the Server service.
type Routes struct {
	*gin.Engine
	DeleteThingsDeviceNameActionsActionNameActionID struct {
		*gin.RouterGroup
	}
	GetThings struct {
		*gin.RouterGroup
	}
	GetThingsDeviceName struct {
		*gin.RouterGroup
	}
	GetThingsDeviceNameActions struct {
		*gin.RouterGroup
	}
	GetThingsDeviceNameActionsActionNameActionID struct {
		*gin.RouterGroup
	}
	GetThingsDeviceNameEvents struct {
		*gin.RouterGroup
	}
	GetThingsDeviceNameEventsEventType struct {
		*gin.RouterGroup
	}
	GetThingsDeviceNameProperties struct {
		*gin.RouterGroup
	}
	GetThingsDeviceNamePropertiesPropertyName struct {
		*gin.RouterGroup
	}
	PostThingsDeviceNameActions struct {
		*gin.RouterGroup
	}
	PutThingsDeviceNameActionsActionNameActionID struct {
		*gin.RouterGroup
	}
	PutThingsDeviceNamePropertiesPropertyName struct {
		*gin.RouterGroup
	}
}

// configureWellKnown enables and configures /.well-known endpoints.
func (r *Routes) configureWellKnown(healthFunc func() bool) {
	wellKnown := r.Group("/.well-known")
	{
		wellKnown.GET("/schema-discovery", func(ctx *gin.Context) {
			discovery := struct {
				SchemaURL  string `json:"schema_url"`
				SchemaType string `json:"schema_type"`
				UIURL      string `json:"ui_url"`
			}{
				SchemaURL:  "/swagger.json",
				SchemaType: "swagger-2.0",
			}
			ctx.JSON(http.StatusOK, &discovery)
		})
		wellKnown.GET("/health", healthHandler(healthFunc))
	}

	r.GET("/swagger.json", func(ctx *gin.Context) {
		ctx.String(http.StatusOK, string(SwaggerJSON))
	})
}

// healthHandler is the health HTTP handler used for the /.well-known/health
// route if enabled.
func healthHandler(healthFunc func() bool) gin.HandlerFunc {
	healthy := healthFunc
	return func(ctx *gin.Context) {
		health := struct {
			Health bool `json:"health"`
		}{
			Health: healthy(),
		}

		if !health.Health {
			ctx.JSON(http.StatusServiceUnavailable, &health)
		} else {
			ctx.JSON(http.StatusOK, &health)
		}
	}
}

// Service is the interface that must be implemented in order to provide
// business logic for the Server service.
type Service interface {
	Healthy() bool
	DeleteThingsDeviceNameActionsActionNameActionID(ctx *gin.Context, params *operations.DeleteThingsDeviceNameActionsActionNameActionIDParams) *api.Response
	GetThings(ctx *gin.Context) *api.Response
	GetThingsDeviceName(ctx *gin.Context, params *operations.GetThingsDeviceNameParams) *api.Response
	GetThingsDeviceNameActions(ctx *gin.Context, params *operations.GetThingsDeviceNameActionsParams) *api.Response
	GetThingsDeviceNameActionsActionNameActionID(ctx *gin.Context, params *operations.GetThingsDeviceNameActionsActionNameActionIDParams) *api.Response
	GetThingsDeviceNameEvents(ctx *gin.Context, params *operations.GetThingsDeviceNameEventsParams) *api.Response
	GetThingsDeviceNameEventsEventType(ctx *gin.Context, params *operations.GetThingsDeviceNameEventsEventTypeParams) *api.Response
	GetThingsDeviceNameProperties(ctx *gin.Context, params *operations.GetThingsDeviceNamePropertiesParams) *api.Response
	GetThingsDeviceNamePropertiesPropertyName(ctx *gin.Context, params *operations.GetThingsDeviceNamePropertiesPropertyNameParams) *api.Response
	PostThingsDeviceNameActions(ctx *gin.Context, params *operations.PostThingsDeviceNameActionsParams) *api.Response
	PutThingsDeviceNameActionsActionNameActionID(ctx *gin.Context, params *operations.PutThingsDeviceNameActionsActionNameActionIDParams) *api.Response
	PutThingsDeviceNamePropertiesPropertyName(ctx *gin.Context, params *operations.PutThingsDeviceNamePropertiesPropertyNameParams) *api.Response
}

func ginizePath(path string) string {
	return strings.Replace(strings.Replace(path, "{", ":", -1), "}", "", -1)
}

// initializeRoutes initializes the route structure for the Server service.
func initializeRoutes(enableAuth bool, tokenURL string, tracer opentracing.Tracer) *Routes {
	engine := gin.New()
	engine.Use(gin.Recovery())
	routes := &Routes{Engine: engine}

	routes.DeleteThingsDeviceNameActionsActionNameActionID.RouterGroup = routes.Group("/v1")
	routes.DeleteThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.DeleteThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(tracing.InitSpan(tracer, "delete_things_device_name_actions_action_name_action_id"))
	}

	routes.GetThings.RouterGroup = routes.Group("/v1")
	routes.GetThings.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThings.RouterGroup.Use(tracing.InitSpan(tracer, "get_things"))
	}

	routes.GetThingsDeviceName.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceName.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceName.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name"))
	}

	routes.GetThingsDeviceNameActions.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceNameActions.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceNameActions.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name_actions"))
	}

	routes.GetThingsDeviceNameActionsActionNameActionID.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name_actions_action_name_action_id"))
	}

	routes.GetThingsDeviceNameEvents.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceNameEvents.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceNameEvents.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name_events"))
	}

	routes.GetThingsDeviceNameEventsEventType.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceNameEventsEventType.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceNameEventsEventType.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name_events_event_type"))
	}

	routes.GetThingsDeviceNameProperties.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceNameProperties.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceNameProperties.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name_properties"))
	}

	routes.GetThingsDeviceNamePropertiesPropertyName.RouterGroup = routes.Group("/v1")
	routes.GetThingsDeviceNamePropertiesPropertyName.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.GetThingsDeviceNamePropertiesPropertyName.RouterGroup.Use(tracing.InitSpan(tracer, "get_things_device_name_properties_property_name"))
	}

	routes.PostThingsDeviceNameActions.RouterGroup = routes.Group("/v1")
	routes.PostThingsDeviceNameActions.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.PostThingsDeviceNameActions.RouterGroup.Use(tracing.InitSpan(tracer, "post_things_device_name_actions"))
	}
	routes.PostThingsDeviceNameActions.RouterGroup.Use(middleware.ContentTypes("application/json"))

	routes.PutThingsDeviceNameActionsActionNameActionID.RouterGroup = routes.Group("/v1")
	routes.PutThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.PutThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(tracing.InitSpan(tracer, "put_things_device_name_actions_action_name_action_id"))
	}
	routes.PutThingsDeviceNameActionsActionNameActionID.RouterGroup.Use(middleware.ContentTypes("application/json"))

	routes.PutThingsDeviceNamePropertiesPropertyName.RouterGroup = routes.Group("/v1")
	routes.PutThingsDeviceNamePropertiesPropertyName.RouterGroup.Use(middleware.LogrusLogger())
	if tracer != nil {
		routes.PutThingsDeviceNamePropertiesPropertyName.RouterGroup.Use(tracing.InitSpan(tracer, "put_things_device_name_properties_property_name"))
	}
	routes.PutThingsDeviceNamePropertiesPropertyName.RouterGroup.Use(middleware.ContentTypes("application/json"))

	return routes
}

// Server defines the Server service.
type Server struct {
	Routes           *Routes
	config           *Config
	server           *http.Server
	service          Service
	healthy          bool
	serviceHealthyFn func() bool
	authDisabled     bool
	Title            string
	Version          string
}

// NewServer initializes a new Server service.
func NewServer(svc Service, config *Config) *Server {
	if !config.Debug {
		gin.SetMode(gin.ReleaseMode)
	}

	server := &Server{
		Routes: initializeRoutes(
			!config.AuthDisabled,
			config.TokenURL,
			config.Tracer,
		),
		service:      svc,
		config:       config,
		Title:        "Web Thing REST API",
		Version:      "0.1.0",
		authDisabled: config.AuthDisabled,
	}

	// enable pprof http endpoints in debug mode
	if config.Debug {
		pprof.Register(server.Routes.Engine)
	}

	// set logrus logger to TextFormatter with no colors
	log.SetFormatter(&log.TextFormatter{DisableColors: true})

	server.server = &http.Server{
		Addr:         config.Address,
		Handler:      server.Routes.Engine,
		ReadTimeout:  10 * time.Second,
		WriteTimeout: 10 * time.Second,
	}

	server.serviceHealthyFn = svc.Healthy

	if !config.WellKnownDisabled {
		server.Routes.configureWellKnown(server.isHealthy)
	}

	// configure healthz endpoint
	server.Routes.GET("/healthz", healthHandler(server.isHealthy))

	return server
}

// isHealthy returns true if both the server and the service reports healthy.
func (s *Server) isHealthy() bool {
	return s.healthy && s.serviceHealthyFn()
}

// ConfigureRoutes starts the internal configureRoutes methode.
func (s *Server) ConfigureRoutes() {
	s.configureRoutes()
}

// configureRoutes configures the routes for the Server service.
// Configuring of routes includes setting up Auth if it is enabled.
func (s *Server) configureRoutes() {
	if !s.authDisabled {
	}

	// setup all service routes after the authenticate middleware has been
	// initialized.
	s.Routes.DeleteThingsDeviceNameActionsActionNameActionID.DELETE(ginizePath("/things/{deviceName}/actions/{actionName}/{actionId}"), operations.DeleteThingsDeviceNameActionsActionNameActionIDEndpoint(s.service.DeleteThingsDeviceNameActionsActionNameActionID))
	s.Routes.GetThings.GET(ginizePath("/things"), operations.GetThingsEndpoint(s.service.GetThings))
	s.Routes.GetThingsDeviceName.GET(ginizePath("/things/{deviceName}"), operations.GetThingsDeviceNameEndpoint(s.service.GetThingsDeviceName))
	s.Routes.GetThingsDeviceNameActions.GET(ginizePath("/things/{deviceName}/actions"), operations.GetThingsDeviceNameActionsEndpoint(s.service.GetThingsDeviceNameActions))
	s.Routes.GetThingsDeviceNameActionsActionNameActionID.GET(ginizePath("/things/{deviceName}/actions/{actionName}/{actionId}"), operations.GetThingsDeviceNameActionsActionNameActionIDEndpoint(s.service.GetThingsDeviceNameActionsActionNameActionID))
	s.Routes.GetThingsDeviceNameEvents.GET(ginizePath("/things/{deviceName}/events"), operations.GetThingsDeviceNameEventsEndpoint(s.service.GetThingsDeviceNameEvents))
	s.Routes.GetThingsDeviceNameEventsEventType.GET(ginizePath("/things/{deviceName}/events/{eventType}"), operations.GetThingsDeviceNameEventsEventTypeEndpoint(s.service.GetThingsDeviceNameEventsEventType))
	s.Routes.GetThingsDeviceNameProperties.GET(ginizePath("/things/{deviceName}/properties"), operations.GetThingsDeviceNamePropertiesEndpoint(s.service.GetThingsDeviceNameProperties))
	s.Routes.GetThingsDeviceNamePropertiesPropertyName.GET(ginizePath("/things/{deviceName}/properties/{propertyName}"), operations.GetThingsDeviceNamePropertiesPropertyNameEndpoint(s.service.GetThingsDeviceNamePropertiesPropertyName))
	s.Routes.PostThingsDeviceNameActions.POST(ginizePath("/things/{deviceName}/actions"), operations.PostThingsDeviceNameActionsEndpoint(s.service.PostThingsDeviceNameActions))
	s.Routes.PutThingsDeviceNameActionsActionNameActionID.PUT(ginizePath("/things/{deviceName}/actions/{actionName}/{actionId}"), operations.PutThingsDeviceNameActionsActionNameActionIDEndpoint(s.service.PutThingsDeviceNameActionsActionNameActionID))
	s.Routes.PutThingsDeviceNamePropertiesPropertyName.PUT(ginizePath("/things/{deviceName}/properties/{propertyName}"), operations.PutThingsDeviceNamePropertiesPropertyNameEndpoint(s.service.PutThingsDeviceNamePropertiesPropertyName))
}

// Run runs the Server. It will listen on either HTTP or HTTPS depending on the
// config passed to NewServer.
func (s *Server) Run() error {
	// configure service routes
	s.configureRoutes()

	log.Infof("Serving '%s - %s' on address %s", s.Title, s.Version, s.server.Addr)
	// server is set to healthy when started.
	s.healthy = true
	if s.config.InsecureHTTP {
		return s.server.ListenAndServe()
	}
	return s.server.ListenAndServeTLS(s.config.TLSCertFile, s.config.TLSKeyFile)
}

// Shutdown will gracefully shutdown the Server server.
func (s *Server) Shutdown() error {
	// server is set to unhealthy when shutting down
	s.healthy = false
	return s.server.Shutdown(context.Background())
}

// RunWithSigHandler runs the Server server with SIGTERM handling automatically
// enabled. The server will listen for a SIGTERM signal and gracefully shutdown
// the web server.
// It's possible to optionally pass any number shutdown functions which will
// execute one by one after the webserver has been shutdown successfully.
func (s *Server) RunWithSigHandler(shutdown ...func() error) error {
	sigCh := make(chan os.Signal, 1)
	signal.Notify(sigCh, syscall.SIGHUP, syscall.SIGINT, syscall.SIGTERM)

	go func() {
		<-sigCh
		s.Shutdown()
	}()

	err := s.Run()
	if err != nil {
		if err != http.ErrServerClosed {
			return err
		}
	}

	for _, fn := range shutdown {
		err := fn()
		if err != nil {
			return err
		}
	}

	return nil
}

// vim: ft=go